HIA-1431: Removed duplicated methods
